Auth: Use SigV4 lib from grafana-aws-sdk  (#30713)
* replace with lib

* remove test + apply feedback
